+++ src/components/QueryEditor/visual-query-builder/SQLGroupByRow.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import { type SelectableValue } from '@grafana/data';
 
 import { GroupByRow } from './GroupByRow';
-import { type QueryWithDefaults } from '../defaults';
-import { type SQLQuery, type DB } from '../types';
+import { type SQLQuery, type DB, type QueryWithDefaults } from '../types';
 import { useSqlChange } from '../utils/useSqlChange';
 
 interface SQLGroupByRowProps {
