+++ src/components/QueryEditor/visual-query-builder/SQLSelectRow.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import { useAsync } from 'react-use';
 import { type SelectableValue, toOption } from '@grafana/data';
 import { SelectRow } from './SelectRow';
-import { type QueryWithDefaults } from '../defaults';
-import { type SQLQuery, type DB } from '../types';
+import { type SQLQuery, type DB, type QueryWithDefaults } from '../types';
 import { useSqlChange } from '../utils/useSqlChange';
 
 interface SQLSelectRowProps {
